requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
flake8==3.9.1
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
